backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymysql==1.1.0
cryptography==41.0.7
django-cors-headers==4.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
factory-boy==3.3.0
coverage==7.3.2
bleach==6.1.0
python-json-logger==2.0.7
requests==2.31.0
